Add assertion in main, also handled wrong inputs
